Fix issue with finding nested partials in Handlebars 1.0.rc.2
This checks the version of handlebars to determine if it's "safe" to use
"/"s in partial names to represent partials logically nested in a file
system. Handlebars prior to version 1.0.rc.2 did not support this, so a
hack was put in to replace "/"s with "."s. This hack is now only applied
to older versions of Handlebars.

Fixes #9