Move file system cache to instance state
This moves the file system cache from static state on the constructor
to instance-based state. This allows for the cache to be garbage
collected when an instance is no longer referenced. It also makes more
sense to scope the cache in this way since the other caches are scoped
to instances.