Solve issue #244
Solve issue #244. When `viewsPath` is truthy, the `partialsDir` which have the directories defined by the user on the constructor, are overwritten by as single String `path.join(viewsPath, 'partials/')`. I'm playing safe here, I assume `partialsDir` is internally always dealt as an Array and therefore I convert it to an Array, pushing when appropriate the default `path.join(viewsPath, 'partials/')`, but keeping the previously defined `partialDirs`